Changed from f32 to f64 to fix shadow rendering 'acne'

+++ src/answers/answers_chapter2.c
@@ -1,25 +1,25 @@
 
 static void answer_chapter2(void) {
-    vec4 velocity = vec4_make_vector(1.0f, 5.0f, 0.0f);
+    vec4 velocity = vec4_make_vector(1.0, 5.0, 0.0);
     velocity      = vec4_normalize(velocity);
-    velocity      = vec4_scalar_mul(velocity, 10.0f);
+    velocity      = vec4_scalar_mul(velocity, 10.0);
 
     projectile proj = {
-        .position = vec4_make_point(0.0f, 250.0f, 0.0f),
+        .position = vec4_make_point(0.0, 250.0, 0.0),
         .velocity = velocity,
     };
     environment env = {
-        .gravity = vec4_make_vector( 0.0f, -0.1f, 0.0f),
-        .wind    = vec4_make_vector(-0.01f, 0.0f, 0.0f),
+        .gravity = vec4_make_vector( 0.0, -0.1, 0.0),
+        .wind    = vec4_make_vector(-0.01, 0.0, 0.0),
     };
 
     canvas *can = canvas_create(250, 800);
-    color3  col = { 0.75f, 0.25f, 0.25f };
-    for (usize i = 1; proj.position.y > 0.0f; i++) {
+    color3  col = { 0.75, 0.25, 0.25 };
+    for (usize i = 1; proj.position.y > 0.0; i++) {
         vec4 pos = proj.position;
 
-        u32 x = (u32) roundf(pos.x);
-        u32 y = (u32) roundf(can->height - 1 - pos.y);
+        u32 x = (u32) round(pos.x);
+        u32 y = (u32) round(can->height - 1 - pos.y);
         canvas_set_pixel(can, x, y, col);
 
         proj = tick(&env, &proj);
