+++ tests/test_renderer_color.c
@@ -1,8 +1,8 @@
 
 void test_add_colors(void) {
-    color3 c1  = { 0.9f, 0.6f, 0.75f };
-    color3 c2  = { 0.7f, 0.1f, 0.25f };
-    color3 sum = { 1.6f, 0.7f, 1.0f  };
+    color3 c1  = { 0.9, 0.6, 0.75 };
+    color3 c2  = { 0.7, 0.1, 0.25 };
+    color3 sum = { 1.6, 0.7, 1.0  };
 
     color3 res = color3_add(c1, c2);
     assert(color3_compare(res, sum) == true);
@@ -10,9 +10,9 @@ void test_add_colors(void) {
 
 
 void test_sub_colors(void) {
-    color3 c1  = { 0.9f, 0.6f, 0.75f };
-    color3 c2  = { 0.7f, 0.1f, 0.25f };
-    color3 sub = { 0.2f, 0.5f, 0.5f  };
+    color3 c1  = { 0.9, 0.6, 0.75 };
+    color3 c2  = { 0.7, 0.1, 0.25 };
+    color3 sub = { 0.2, 0.5, 0.5  };
 
     color3 res = color3_sub(c1, c2);
     assert(color3_compare(res, sub) == true);
@@ -20,9 +20,9 @@ void test_sub_colors(void) {
 
 
 void test_scalar_mul_color(void) {
-    color3 c      = { 0.2f, 0.3f, 0.4f };
-    color3 mul    = { 0.4f, 0.6f, 0.8f };
-    f32    scalar = 2.0f;
+    color3 c      = { 0.2, 0.3, 0.4 };
+    color3 mul    = { 0.4, 0.6, 0.8 };
+    f64    scalar = 2.0;
 
     color3 res = color3_scalar_mul(c, scalar);
     assert(color3_compare(res, mul) == true);
@@ -30,9 +30,9 @@ void test_scalar_mul_color(void) {
 
 
 void test_hadamard_product_color(void) {
-    color3 c1  = { 1.0f, 0.2f, 0.4f  };
-    color3 c2  = { 0.9f, 1.0f, 0.1f  };
-    color3 mul = { 0.9f, 0.2f, 0.04f };
+    color3 c1  = { 1.0, 0.2, 0.4  };
+    color3 c2  = { 0.9, 1.0, 0.1  };
+    color3 mul = { 0.9, 0.2, 0.04 };
 
     color3 res = color3_mul(c1, c2);
     assert(color3_compare(res, mul) == true);
